requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
datasets==3.6.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
executing==2.2.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
importlib_metadata==8.7.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaleido==1.0.0
Levenshtein==0.27.1
litellm==1.72.7
llvmlite==0.44.0
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
narwhals==1.43.1
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.90.0
orjson==3.10.18
packaging==25.0
pandas==2.3.0
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
sentry-sdk==2.30.0
setproctitle==1.3.6
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.5.0
wandb==0.20.1
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
nbformat==5.10.4

